Add only update repository if version number has changed

+++ update_repo.py
@@ -29,27 +29,28 @@ if __name__ == "__main__":
                 shutil.copy(source_addon_xml, target_addon_xml)
                 if os.path.exists(source_addon_icon):
                     shutil.copy(source_addon_icon, target_addon_icon)
-                if addon_name.startswith('repository.'):
-                    if os.path.exists(target_addon_zip):
-                        os.unlink(target_addon_zip)
-                    source_addon_dir = addon_name
-                with zipfile.ZipFile(target_addon_zip, 'w', zipfile.ZIP_DEFLATED) as zip:
-                    for root, dirs, files in os.walk(source_addon_dir):
-                        for f in files:
-                            filename = os.path.join(root, f)
-                            if source_addon_dir == addon_name:
-                                arcname = filename
-                            else:
-                                arcname = os.path.relpath(filename, os.path.join('..'))
+                if not os.path.exists(target_addon_zip):
+                    if addon_name.startswith('repository.'):
+                        if os.path.exists(target_addon_zip):
+                            os.unlink(target_addon_zip)
+                        source_addon_dir = addon_name
+                    with zipfile.ZipFile(target_addon_zip, 'w', zipfile.ZIP_DEFLATED) as zip:
+                        for root, dirs, files in os.walk(source_addon_dir):
+                            for f in files:
+                                filename = os.path.join(root, f)
+                                if source_addon_dir == addon_name:
+                                    arcname = filename
+                                else:
+                                    arcname = os.path.relpath(filename, os.path.join('..'))
 
-                            if f in [
-                                'addon.py',
-                                'addon.xml',
-                                'changelog.txt',
-                                'fanart.jpg',
-                                'icon.png',
-                                'LICENSE.txt'
-                            ] or arcname.startswith(os.path.join(addon_name, 'resources')):
-                                zip.write(filename, arcname)
+                                if f in [
+                                    'addon.py',
+                                    'addon.xml',
+                                    'changelog.txt',
+                                    'fanart.jpg',
+                                    'icon.png',
+                                    'LICENSE.txt'
+                                ] or arcname.startswith(os.path.join(addon_name, 'resources')):
+                                    zip.write(filename, arcname)
 
         Generator()
